requirements.txt
asgiref==3.5.2
charset-normalizer==2.1.1
Django==3.2.14
django-bootstrap-modal-forms==2.2.0
django-braces==1.15.0
django-ckeditor==6.4.2
django-debug-toolbar==3.5.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==5.0
django-simple-captcha==0.5.17
django-star-ratings==0.9.2
django-taggit==3.0.0
django-widget-tweaks==1.4.12
mysqlclient==2.1.1
Pillow==9.2.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
swapper==1.3.0
tzdata==2022.1
